Basic action for fetching or building jruby-head
This is used within JRuby's GHA builds, to reuse the common build
artifacts most jobs use. The action will attempt to download a
cached build for this hash, and failing that it will do a full
build and upload the result. This means jobs that start
immediately will all try to be the first to build and upload, but
none will be blocked on some other job building and uploading
first.